refactor: fixed videos page for rc10
