+++ package.json
@@ -30,7 +30,7 @@
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.7.0",
+    "@headlessui/vue": "^1.7.2",
     "@iconify/vue": "^4.0.0",
     "@nuxtjs/algolia": "^1.1.1",
     "@pinia/nuxt": "^0.4.2",
@@ -50,8 +50,8 @@
     "@histoire/plugin-nuxt": "^0.10.7",
     "@intlify/eslint-plugin-vue-i18n": "^2.0.0",
     "@intlify/nuxt3": "^0.2.4",
-    "@nuxt/content": "^2.1.0",
-    "@nuxt/test-utils": "3.0.0-rc.9",
+    "@nuxt/content": "^2.1.1",
+    "@nuxt/test-utils": "3.0.0-rc.10",
     "@nuxtjs/tailwindcss": "^5.3.3",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
@@ -74,7 +74,7 @@
     "tailwind-scrollbar-hide": "^1.1.7",
     "tailwindcss": "^3.1.8",
     "typescript": "^4.8.3",
-    "vitest": "^0.23.1"
+    "vitest": "^0.23.2"
   },
   "lint-staged": {
     "*.{js,ts}": "eslint --cache --fix"
