+++ pages/videos/index.vue
@@ -4,8 +4,10 @@
       <h1 class="text-center text-6xl font-bold">
         {{ t('page.videos.head.title') }}
       </h1>
-      <ActiveVideosHorizontalFilter />
-      <div class="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-3">
+      <ActiveVideosHorizontalFilter/>
+      <div
+        v-if="videos.length"
+        class="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-3">
         <MoleculesCardVideo
           v-for="video in videos"
           :key="video._path"
@@ -23,6 +25,7 @@
         />
       </div>
       <MoleculesPagination
+        v-if="pages?.total > 1"
         :total="pages.total"
         :number-of-elements="8"
         :current-page="pages.current"
@@ -32,14 +35,18 @@
 </template>
 
 <script setup lang="ts">
-  import { useVideosList } from '~/store/videos/videoList';
-  import { useI18n } from 'vue-i18n';
+import {useI18n} from 'vue-i18n';
+import {useVideosList} from '~/composables/store/useVideosList';
 
-  definePageMeta({
-    title: 'i18n:page.videos.head.title',
-  });
+definePageMeta({
+  title: 'i18n:page.videos.head.title',
+});
 
-  const { t } = useI18n();
+const {t} = useI18n();
 
-  const { pages, videos } = useVideosList();
+const {
+  data: videos,
+  filters,
+  pages,
+} = useVideosList();
 </script>
