+++ types/components/filter.ts
@@ -1,6 +1,8 @@
+import {LocationQueryValue} from 'vue-router';
+
 export type Filter = {
   id: string;
   name: string;
   options: string[];
-  selected: Set<string>;
+  selected: Set<LocationQueryValue | LocationQueryValue[]>;
 };
